Added CreateSchema function to utils lib and configured generate component file
