+++ lib/utils.js
@@ -15,4 +15,25 @@ function createDir(location, name) {
   }
 }
 
-module.exports = { createDir };
+/**
+ * @param {string} locationInSchemasTemplates
+ * @param {string} locationToPaste
+ * @param {(content: string) => string} customize
+ */
+function createSchema(locationInSchemasTemplates, locationToPaste, customize) {
+  const templatePath = `${__dirname}/../templates/schemas/${locationInSchemasTemplates}`;
+  const filesToCreate = fs.readdirSync(templatePath);
+
+  filesToCreate.forEach((file) => {
+    const origFilePath = `${templatePath}/${file}`;
+    let contents = fs.readFileSync(origFilePath, 'utf8');
+
+    const writePath = `${locationToPaste}/${file}`;
+    if (customize) {
+      contents = customize(contents);
+    }
+    fs.writeFileSync(writePath, contents, 'utf8');
+  });
+}
+
+module.exports = { createDir, createSchema };
